refactor: mv GitChangelogOptions types
