+++ packages/vitepress-plugin-git-changelog/src/vite/git.ts
@@ -8,7 +8,6 @@ import { execa } from 'execa'
 import type { Changelog, Commit } from '../types'
 import {
   type CommitToStringHandler,
-  type CommitToStringsHandler,
   type RewritePathsBy,
   defaultCommitURLHandler,
   defaultReleaseTagURLHandler,
@@ -16,6 +15,7 @@ import {
   getCommits,
   rewritePathsByRewritingExtension,
 } from './helpers'
+import type { GitChangelogOptions } from './types'
 
 export type {
   CommitToStringHandler,
@@ -30,117 +30,6 @@ const VirtualModuleID = 'virtual:nolebase-git-changelog'
 const ResolvedVirtualModuleId = `\0${VirtualModuleID}`
 
 const logModulePrefix = `${cyan(`@nolebase/vitepress-plugin-git-changelog`)}${gray(':')}`
-
-export interface GitChangelogOptions {
-  /**
-   * The current working directory in which to search.
-   *
-   * @default process.cwd()
-   */
-  cwd?: string
-  /**
-   * When fetching git logs, what files should be included?
-   *
-   * @default ['** /*.md', '!node_modules']
-   */
-  include?: string[]
-  /**
-   * When fetching git logs, what directories should be included?
-   *
-   * @deprecated
-   */
-  includeDirs?: string[]
-  /**
-   * When fetching git logs, what extensions should be included?
-   *
-   * @deprecated
-   */
-  includeExtensions?: `.${string}`[]
-  /**
-   * Your repository URL.
-   * Yes, you can dynamically generate it.
-   *
-   * @default 'https://github.com/example/example'
-   */
-  repoURL?: string | CommitToStringHandler
-  /**
-   * A function to get the release tag URL.
-   *
-   * @default (commit) => `${commit.repo_url}/releases/tag/${commit.tag}`
-   */
-  getReleaseTagURL?: CommitToStringHandler
-  /**
-   * A function to get the release tags URL.
-   *
-   * @default (commit) => commit.tags?.map(tag => `${commit.repo_url}/releases/tag/${tag}`)
-   */
-  getReleaseTagsURL?: CommitToStringsHandler
-  /**
-   * A function to get the commit URL.
-   *
-   * @default (commit) => `${commit.repo_url}/commit/${commit.hash}`
-   */
-  getCommitURL?: CommitToStringHandler
-  /**
-   * A map that contains rewrite rules of paths.
-   *
-   * This is quite useful when you have your pages in a different directory than the base url after deployed since the
-   * data will be calculated again on the client side.
-   *
-   * For example:
-   *  - We have a page at `docs/pages/en/integrations/page.md`
-   *  - And we will deploy it to `https://example.com/en/integrations/page`
-   *
-   * Then you can set the rewrite paths like this:
-   * ```json
-   * {
-   *  "docs/": ""
-   * }
-   * ```
-   *
-   * This will rewrite the path to `en/integrations/page.md`
-   * Which is the correct path for the deployed page and runtime scripts to work properly.
-   *
-   * Note: in runtime, which is client side, the final extension will be replaced with `.md` if the extension is `.html`.
-   *
-   */
-  rewritePaths?: Record<string, string>
-  /**
-   * Rules to rewrite paths by patterns.
-   *
-   * Same as `rewritePaths`, but it can be a function that returns a promise or plain value.
-   * Besides that, we offer some built-in handlers to rewrite paths by patterns:
-   *
-   *  - `rewritePathsByRewritingExtension(from: string, to: string)`: to rewrite paths by rewriting the extension.
-   *
-   * @example
-   *
-   * ```typescript
-   * import { GitChangelog, rewritePathsByRewritingExtension } from '@nolebase/vitepress-plugin-git-changelog/vite'
-   *
-   * GitChangelog({
-   *  rewritePathsBy: {
-   *   // to rewrite `example.md` to `example.html`
-   *  handler: rewritePathsByRewritingExtension('.md', '.html')
-   * }
-   * })
-   * ```
-   *
-   * @see rewritePathsByRewritingExtension
-   *
-   */
-  rewritePathsBy?: RewritePathsBy
-  /**
-   * The maximum number of git logs to fetch.
-   */
-  maxGitLogCount?: number
-  /**
-   * The maximum number of concurrent processes to fetch git logs.
-   *
-   * @deprecated
-   */
-  maxConcurrentProcesses?: number
-}
 
 export function GitChangelog(options: GitChangelogOptions = {}): Plugin {
   if (!options)
